pyproject.toml
[project]
name = "snapload-api"
version = "1.0.0"
description = "SnapLoad API - REST API for downloading media from YouTube and Spotify"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.1",
    "fastapi>=0.103.2",
    "python-dotenv>=1.2.1",
    "spotdl>=4.4.3",
    "uvicorn>=0.23.2",
    "yt-dlp>=2025.11.12",
]
